Update 14 - 11/23/2024 - 11:48 PM - JB
Went a little crazy with the changes. This update represents an entire day spent doing nothing but working on this project. Decided to be done with Irvine, it was way to much of a hasle. Finally added a win state. Split the driving process in controller.asm into multiple processes to make it easier to keep jumps straight. Decided to take the return address off of the stack for any process that loops back to start a new round and have the program end by calling a exitProgram in main.asm instead of going back through the recursive return addresses to get back to the start. Ran out of time to work on this for today, but the processes for capturing a pit are still a bit broken, and I was in the middle of adding a reference to board.asm from the postProcessing in controller.asm to get the correct pit that was captured. Need to add a new process in board.asm to get that number.